genome decoding

+++ src/main/java/genetics/genes/NodeGene.java
@@ -3,14 +3,14 @@ package genetics.genes;
 public class NodeGene extends Gene {
 
     public int id;
-    public int bias;
+    public float bias;
     public NodeType nodeType;
 
     public enum NodeType {
         INPUT, HIDDEN, OUTPUT
     }
 
-    public NodeGene(int id, int bias, NodeType nodeType) {
+    public NodeGene(int id, float bias, NodeType nodeType) {
         this.id = id;
         this.bias = bias;
         this.nodeType = nodeType;
